# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CRM for repair orders management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: Get all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get orders from buffer
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderRequest"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/assign-master:
    post:
      operationId: assignMaster
      tags: [orders]
      summary: Assign master to order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignMasterRequest"
      responses:
        "200":
          description: Master assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /settings/cities:
    get:
      operationId: getCities
      tags: [settings]
      summary: Get cities
      responses:
        "200":
          description: List of cities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/City"
    post:
      operationId: createCity
      tags: [settings]
      summary: Create city
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCityRequest"
      responses:
        "201":
          description: City created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/City"
  /settings/cities/{id}:
    delete:
      operationId: deleteCity
      tags: [settings]
      summary: Delete city
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: City deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /settings/services:
    get:
      operationId: getServices
      tags: [settings]
      summary: Get service types
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceType"
    post:
      operationId: createService
      tags: [settings]
      summary: Create service type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceType"
  /settings/services/{id}:
    delete:
      operationId: deleteService
      tags: [settings]
      summary: Delete service type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    User:
      type: object
      properties:
        id:
          type: integer
        login:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, lead_operator, master_operator]
        createdAt:
          type: string
          format: date-time
      required: [id, login, name, role, createdAt]
    CreateUserRequest:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, lead_operator, master_operator]
      required: [login, password, name, role]
...
    Order:
      type: object
      properties:
        id:
          type: integer
        leadId:
          type: integer
        city:
          type: string
        district:
          type: string
        serviceType:
          type: string
        area:
          type: number
        scheduledAt:
          type: string
          nullable: true
        comment:
          type: string
          nullable: true
        status:
          type: string
          enum: [waiting_master, master_assigned, in_progress, completed, cancelled]
        masterId:
          type: integer
          nullable: true
        masterName:
          type: string
          nullable: true
        orderAmount:
          type: number
          nullable: true
        commission:
          type: number
          nullable: true
        clientRating:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, leadId, city, district, serviceType, area, status, createdAt, updatedAt]
    UpdateOrderRequest:
      type: object
      properties:
        status:
          type: string
          enum: [waiting_master, master_assigned, in_progress, completed, cancelled]
        orderAmount:
          type: number
          nullable: true
        commission:
          type: number
          nullable: true
        clientRating:
          type: integer
          nullable: true
    AssignMasterRequest:
      type: object
      properties:
        masterId:
          type: integer
      required: [masterId]
...
    City:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required: [id, name]
    CreateCityRequest:
      type: object
      properties:
        name:
          type: string
      required: [name]
    ServiceType:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required: [id, name]
    CreateServiceRequest:
      type: object
      properties:
        name:
          type: string
      required: [name]