# lib/api-zod/src/generated/types/createServiceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CRM for repair orders management
 * OpenAPI spec version: 0.1.0
 */

export interface CreateServiceRequest {
  name: string;
}