# lib/api-zod/src/generated/types/createUserRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CRM for repair orders management
 * OpenAPI spec version: 0.1.0
 */
import type { CreateUserRequestRole } from "./createUserRequestRole";

export interface CreateUserRequest {
  login: string;
  password: string;
  name: string;
  role: CreateUserRequestRole;
}