# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CRM for repair orders management
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  leadId: number;
  city: string;
  district: string;
  serviceType: string;
  area: number;
  scheduledAt?: string | null;
  comment?: string | null;
  status: OrderStatus;
  masterId?: number | null;
  masterName?: string | null;
  orderAmount?: number | null;
  commission?: number | null;
  clientRating?: number | null;
  createdAt: Date;
  updatedAt: Date;
}